requirements.txt
asgiref==3.8.1
Django==4.2.13
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
importlib_metadata==7.1.0
Markdown==3.6
PyJWT==2.8.0
sqlparse==0.5.0
typing_extensions==4.12.2
zipp==3.19.2
